Module m3u8_rs::playlist

Contains all the structs used for parsing.

The main type here is the Playlist enum. Which is either a MasterPlaylist or a MediaPlaylist.

Structs

AlternativeMedia
#EXT-X-MEDIA:<attribute-list>

ByteRange
#EXT-X-BYTERANGE:<n>[@<o>]

DateRange
#EXT-X-DATERANGE:<attribute-list>

ExtTag
A simple #EXT- tag

Key
#EXT-X-KEY:<attribute-list>

Map
#EXT-X-MAP:<attribute-list>

MasterPlaylist
A Master Playlist provides a set of Variant Streams, each of which describes a different version of the same content.

MediaPlaylist
A Media Playlist contains a list of Media Segments, which when played sequentially will play the multimedia presentation.

MediaSegment
A Media Segment is specified by a URI and optionally a byte range.

SessionData
#EXT-X-SESSION-DATA:<attribute-list> The EXT-X-SESSION-KEY tag allows encryption keys from Media Playlists to be specified in a Master Playlist. This allows the client to preload these keys without having to read the Media Playlist(s) first.

SessionKey
#EXT-X-SESSION-KEY:<attribute-list>

Start
#EXT-X-START:<attribute-list>

VariantStream
#EXT-X-STREAM-INF:<attribute-list> <URI> #EXT-X-I-FRAME-STREAM-INF:<attribute-list>

Enums

AlternativeMediaType

MediaPlaylistType
#EXT-X-PLAYLIST-TYPE:<EVENT|VOD>

Playlist
Playlist, can either be a MasterPlaylist or a MediaPlaylist.