Enum m3u8_rs::playlist::Playlist

pub enum Playlist {
    MasterPlaylist(MasterPlaylist),
    MediaPlaylist(MediaPlaylist),
}

Playlist, can either be a MasterPlaylist or a MediaPlaylist.

A Playlist is a Media Playlist if all URI lines in the Playlist identify Media Segments. A Playlist is a Master Playlist if all URI lines in the Playlist identify Media Playlists. A Playlist MUST be either a Media Playlist or a Master Playlist; all other Playlists are invalid.

Variants

MasterPlaylist

MediaPlaylist