Struct m3u8_rs::playlist::AlternativeMedia

pub struct AlternativeMedia {
    pub media_type: AlternativeMediaType,
    pub uri: Option<String>,
    pub group_id: String,
    pub language: Option<String>,
    pub assoc_language: Option<String>,
    pub name: String,
    pub default: bool,
    pub autoselect: bool,
    pub forced: bool,
    pub instream_id: Option<String>,
    pub characteristics: Option<String>,
}

#EXT-X-MEDIA:<attribute-list>

The EXT-X-MEDIA tag is used to relate Media Playlists that contain alternative Renditions (Section 4.3.4.2.1) of the same content. For example, three EXT-X-MEDIA tags can be used to identify audio-only Media Playlists that contain English, French and Spanish Renditions of the same presentation. Or two EXT-X-MEDIA tags can be used to identify video-only Media Playlists that show two different camera angles.

Fields
media_type: AlternativeMediaType uri: Option<String> group_id: String language: Option<String> assoc_language: Option<String> name: String default: bool autoselect: bool forced: bool instream_id: Option<String> characteristics: Option<String>
Methods

impl AlternativeMedia

fn from_hashmap(attrs: HashMap<String, String>) -> AlternativeMedia

fn write_to<T: Write>(&self, w: &mut T) -> Result<()>