Struct m3u8_rs::playlist::MasterPlaylist

pub struct MasterPlaylist {
    pub version: usize,
    pub variants: Vec<VariantStream>,
    pub session_data: Option<SessionData>,
    pub session_key: Option<SessionKey>,
    pub start: Option<Start>,
    pub independent_segments: bool,
}

A Master Playlist provides a set of Variant Streams, each of which describes a different version of the same content.

Fields

version

variants

session_data

session_key

start

independent_segments

Methods

impl MasterPlaylist

fn from_tags(tags: Vec<MasterPlaylistTag>) -> MasterPlaylist