Struct m3u8_rs::playlist::MediaPlaylist

pub struct MediaPlaylist {
    pub version: usize,
    pub target_duration: f32,
    pub media_sequence: i32,
    pub segments: Vec<MediaSegment>,
    pub discontinuity_sequence: i32,
    pub end_list: bool,
    pub playlist_type: Option<MediaPlaylistType>,
    pub i_frames_only: bool,
    pub start: Option<Start>,
    pub independent_segments: bool,
}

A Media Playlist contains a list of Media Segments, which when played sequentially will play the multimedia presentation.

Fields
version: usize target_duration: f32

#EXT-X-TARGETDURATION:<s>

media_sequence: i32

#EXT-X-MEDIA-SEQUENCE:<number>

segments: Vec<MediaSegment> discontinuity_sequence: i32

#EXT-X-DISCONTINUITY-SEQUENCE:<number>

end_list: bool

#EXT-X-ENDLIST

playlist_type: Option<MediaPlaylistType>

#EXT-X-PLAYLIST-TYPE

i_frames_only: bool

#EXT-X-I-FRAMES-ONLY

start: Option<Start>

#EXT-X-START

independent_segments: bool

#EXT-X-INDEPENDENT-SEGMENTS

Methods

impl MediaPlaylist

fn from_tags(tags: Vec<MediaPlaylistTag>) -> MediaPlaylist

fn write_to<T: Write>(&self, w: &mut T) -> Result<()>