Struct m3u8_rs::playlist::SessionData

pub struct SessionData {
    pub data_id: String,
    pub value: String,
    pub uri: String,
    pub language: Option<String>,
}

#EXT-X-SESSION-DATA:<attribute-list> The EXT-X-SESSION-KEY tag allows encryption keys from Media Playlists to be specified in a Master Playlist. This allows the client to preload these keys without having to read the Media Playlist(s) first.

Fields

data_id

value

uri

language

Methods

impl SessionData

fn from_hashmap(attrs: HashMap<String, String>) -> SessionData