Struct m3u8_rs::playlist::VariantStream

pub struct VariantStream {
    pub is_i_frame: bool,
    pub uri: String,
    pub bandwidth: String,
    pub average_bandwidth: Option<String>,
    pub codecs: String,
    pub resolution: Option<String>,
    pub frame_rate: Option<String>,
    pub audio: Option<String>,
    pub video: Option<String>,
    pub subtitles: Option<String>,
    pub closed_captions: Option<String>,
    pub alternatives: Vec<AlternativeMedia>,
}

#EXT-X-STREAM-INF:<attribute-list> #EXT-X-I-FRAME-STREAM-INF:<attribute-list>

A Variant Stream includes a Media Playlist that specifies media encoded at a particular bit rate, in a particular format, and at a particular resolution for media containing video.

A Variant Stream can also specify a set of Renditions. Renditions are alternate versions of the content, such as audio produced in different languages or video recorded from different camera angles.

Clients should switch between different Variant Streams to adapt to network conditions. Clients should choose Renditions based on user preferences.

Fields

is_i_frame

uri

bandwidth

average_bandwidth

codecs

resolution

frame_rate

audio

video

subtitles

closed_captions

alternatives

Methods

impl VariantStream

fn from_hashmap(attrs: HashMap<String, String>, is_i_frame: bool) -> VariantStream