Struct m3u8_rs::playlist::MediaPlaylist

pub struct MediaPlaylist {
    pub version: usize,
    pub target_duration: f32,
    pub media_sequence: i32,
    pub segments: Vec<MediaSegment>,
    pub discontinuity_sequence: i32,
    pub end_list: bool,
    pub playlist_type: MediaPlaylistType,
    pub i_frames_only: bool,
    pub start: Option<Start>,
    pub independent_segments: bool,
}

A Media Playlist contains a list of Media Segments, which when played sequentially will play the multimedia presentation.

Fields

version

target_duration

#EXT-X-TARGETDURATION:<s>

media_sequence

#EXT-X-MEDIA-SEQUENCE:<number>

segments

discontinuity_sequence

#EXT-X-DISCONTINUITY-SEQUENCE:<number>

end_list

#EXT-X-ENDLIST

playlist_type

#EXT-X-PLAYLIST-TYPE

i_frames_only

#EXT-X-I-FRAMES-ONLY

start

#EXT-X-START

independent_segments

#EXT-X-INDEPENDENT-SEGMENTS

Methods

impl MediaPlaylist

fn from_tags(tags: Vec<MediaPlaylistTag>) -> MediaPlaylist